requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
biopython>=1.79
scikit-learn>=0.24.0
networkx>=2.6.0
tqdm>=4.62.0 